Pre-allocation and caching
In CPython, the first time a function is called, it is compiled into bytecode and a number of small integers
are pre-allocated and cached for performance reasons.
This means that after the first line of code print("I") is executed,
262 int objects would be created and cached by the Python interpreter.
These int objects correspond to the range of small integers defined by NSMALLPOSINTS and NSMALLNEGINTS,
which by default are set to 257 and 5 respectively,
meaning that integers from -5 to 256 are pre-allocated and cached.